updated license, GPL fuck off!!!!!!!!!!!!!!!!!!!!!!!!!!!
I have strong reservations about the GPL. While I appreciate the principles of copyleft, I find the limitations of GPL versions 1, 2, and 3 to be restrictive. If you're interested in using copyleft, I encourage you to choose a copyleft license that promotes collaboration and allows for contributions to be shared freely under previous licenses.